Rename syncer/controllers to syncer/resources
