+++ incubator/virtualcluster/pkg/syncer/resources/pod/checker.go
